% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_point_scramble.R
\name{compute_group_scramble}
\alias{compute_group_scramble}
\title{Title}
\usage{
compute_group_scramble(data, scales, seed = NULL)
}
\arguments{
\item{seed}{}
}
\value{

}
\description{
Title
}
\examples{
cars \%>\%
rename(x = speed, y = dist) \%>\%
compute_group_scramble()
}
