% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_sample_prop.R
\name{facet_sample_prop}
\alias{facet_sample_prop}
\title{Title}
\usage{
facet_sample_prop(
  n = 9,
  prop = 0.2,
  nrow = NULL,
  ncol = NULL,
  scales = "fixed",
  shrink = TRUE,
  strip.position = "top",
  seed = sample(2000:3000, 1)
)
}
\arguments{
\item{n}{number of facets}

\item{prop}{numeric proportion to sample, defaults to .2}

\item{nrow}{how many rows for facet grid}

\item{ncol}{how many columns for facet grid}

\item{scales}{should x and y scales be determined independently for facets}

\item{shrink}{not sure what shrink does}

\item{strip.position}{where facet label should go}

\item{seed}{randomization start point}
}
\value{

}
\description{
Title
}
\examples{
set.seed(1323)
library(ggplot2)
ggplot(data = cars) +
  aes(x = speed) +
  aes(y = dist) +
  geom_point(color = "olivedrab4") +
  geom_smooth(method = lm, se = FALSE) +
  facet_sample_prop() +
  labs(title = "Random Sample, 20\% of population")
}
