% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_scramble.R
\name{facet_scramble}
\alias{facet_scramble}
\title{Title}
\usage{
facet_scramble(
  n_facets = 9,
  prop = 1,
  nrow = NULL,
  ncol = NULL,
  scales = "fixed",
  shrink = TRUE,
  strip.position = "top",
  seed = sample(2000:3000, 1)
)
}
\arguments{
\item{n_facets}{number of facets}

\item{prop}{numeric proportion to sample, defaults to .2}

\item{nrow}{how many rows for facet grid}

\item{ncol}{how many columns for facet grid}

\item{scales}{should x and y scales be determined independently for facets}

\item{shrink}{not sure what shrink does}

\item{strip.position}{where facet label should go}

\item{seed}{randomization start point}
}
\value{

}
\description{
Title
}
\examples{
library(ggplot2)
set.seed(232)
ggplot(data = cars) +
  aes(x = speed) +
  aes(y = dist) +
  geom_rug() +
  geom_point(color = "springgreen4",
           alpha = .75) +
  facet_scramble(n_facets = 1) +
  facet_scramble(n_facets = 2) +
  facet_scramble(n_facets = 3) +
  ggxmean::geom_lm() +
  facet_scramble(n_facets = 12) +
  labs(title = "Disassociating variables")

}
