% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_lm_sample.R
\name{geom_lm_sample}
\alias{geom_lm_sample}
\title{A geom that shows the linear model of a sample of observations}
\usage{
geom_lm_sample(
  mapping = NULL,
  data = NULL,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{aesthetic mappings}

\item{data}{global data}

\item{position}{ggplot2 positions}

\item{na.rm}{logical remove na}

\item{show.legend}{logical, should legend be displayed for layer}

\item{inherit.aes}{logical, should global aes be used for this layer}

\item{...}{all the geom_point arguments}
}
\value{
a geom layer
}
\description{
A geom that shows the linear model of a sample of observations
}
\examples{
my_seed <- sample(1:100, 1)
library(ggplot2)
ggplot(data = cars) +
  aes(x = speed, y = dist) +
  geom_point() +
  geom_smooth(method = lm,
             se = FALSE) +
   geom_point_sample(color = "red",
                     seed = my_seed) +
   geom_lm_sample(color = "red",
                  seed = my_seed)

}
