% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_point_lm_sample.R
\name{geom_point_lm_sample}
\alias{geom_point_lm_sample}
\title{A geom that shows the linear model of a sample of observations}
\usage{
geom_point_lm_sample(
  sample_size = 10,
  seed = sample(3000:4000, 1),
  color = "blue"
)
}
\value{
a geom layer
}
\description{
A geom that shows the linear model of a sample of observations
}
\examples{
my_seed <- sample(1:100, 1)
library(ggplot2)
ggplot(data = cars) +
  aes(x = speed, y = dist) +
  geom_point() +
  geom_point_lm_sample() +
  geom_point_lm_sample(color = "darkred")
}
