% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_point_sample.R
\name{geom_point_sample}
\alias{geom_point_sample}
\title{A geom that shows a sampling of points}
\usage{
geom_point_sample(
  mapping = NULL,
  data = NULL,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{aesthetic mappings}

\item{data}{global data}

\item{position}{ggplot2 positions}

\item{na.rm}{logical remove na}

\item{show.legend}{logical, should legend be displayed for layer}

\item{inherit.aes}{logical, should global aes be used for this layer}

\item{...}{all the geom_point arguments}
}
\value{
ggplot2 layer
}
\description{
A geom that shows a sampling of points
}
\examples{
# using the function
library(ggplot2)
ggplot(data = cars) +
 aes(x = speed, y = dist) +
 geom_point() +
 geom_point_sample(color = "red") +
 geom_point_sample(color = "green", sample_size = 5) +
 geom_point_sample(sample_size = 2, color = "blue")

}
