% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_sample.R
\name{geom_triangle_point}
\alias{geom_triangle_point}
\title{Triangular shape points for displaying scatterplot}
\usage{
geom_triangle_point(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created
by aes or aes_. If specified and inherit.aes = TRUE
(the default), is combined with the default mapping at
the top level of the plot. You only need to supply
mapping if there isn't a mapping defined for the plot.}

\item{data}{data frame. If specified, overrides the default
data frame defined at the top level of the plot.}

\item{stat}{The statistical transformation to use on
the data for this layer, as a string.}

\item{position}{Position adjustment, either as a string,
or the result of a call to a position adjustment function.}

\item{na.rm}{f FALSE (the default), removes missing values
with a warning. If TRUE silently removes missing values.}

\item{show.legend}{logical. Should this layer be included in
the legends? NA, the default, includes if any aesthetics are
mapped. FALSE never includes, and TRUE always includes.}

\item{inherit.aes}{If FALSE, overrides the default aesthetics,
rather than combining with them. This is most useful for helper
functions that define both data and aesthetics and shouldn't inherit
behaviour from the default plot specification}

\item{...}{other arguments passed on to layer. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{color = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat}
}
\description{
The triangle point (triangle shape) geom is used to create scatterplot
}
\details{
The scatterplot is useful for displaying the relationship between two
continuous variables.
}
\examples{
p <- ggplot(cars, aes(speed, dist))
p + geom_triangle_point()

}

