% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_sample.R
\name{facet_sample}
\alias{facet_sample}
\title{Title}
\usage{
facet_sample(
  n = 16,
  n_sampled = 5,
  nrow = NULL,
  ncol = NULL,
  scales = "fixed",
  shrink = TRUE,
  strip.position = "top",
  seed = sample(2000:3000, 1)
)
}
\arguments{
\item{n}{number of facets}

\item{n_sampled}{sample size}

\item{nrow}{how many rows for facet grid}

\item{ncol}{how many columns for facet grid}

\item{scales}{should x and y scales be determined independently for facets}

\item{shrink}{not sure what shrink does}

\item{strip.position}{where facet label should go}

\item{seed}{randomization start point}
}
\value{
a ggplot2 facet specification
}
\description{
Title
}
\examples{
set.seed(1323)
library(ggplot2)
ggplot(data = cars) +
  aes(x = speed) +
  aes(y = dist) +
  geom_point(color = "olivedrab4") +
  geom_smooth(method = lm,
              se = FALSE) +
  labs(caption = "population are observations from cars correlation study dataset") +
  facet_sample(n_sampled = 8) +
  labs(title = "16 draws, random sample of 8 observations from the population") +
  labs(subtitle = "Depending on our particular random sample, slopes are flatter or steeper\nand intercepts vary")
}
